+++ tests/test_accesslog.py
@@ -1,14 +1,18 @@
 import json
 import pathlib
+import datetime
 
 import pytest
 from _pytest.capture import CaptureFixture
+from _pytest.fixtures import FixtureRequest
 
 from spinta.accesslog.file import FileAccessLog
+from spinta.auth import get_default_auth_client_id, load_key, KeyType, create_access_token
 from spinta.components import Store
 from spinta.core.config import RawConfig
 from spinta.testing.client import create_test_client
 from spinta.testing.context import create_test_context
+from spinta.testing.manifest import bootstrap_manifest
 
 
 def _upload_pdf(model, app):
@@ -43,6 +47,7 @@ def test_post_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'client': 'test-client',
@@ -85,6 +90,7 @@ def test_post_array_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'method': 'POST',
@@ -132,6 +138,7 @@ def test_put_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'method': 'PUT',
@@ -170,6 +177,7 @@ def test_pdf_put_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'action': 'update',
@@ -219,6 +227,7 @@ def test_patch_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -263,6 +272,7 @@ def test_get_accesslog(app, model, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -308,6 +318,7 @@ def test_get_array_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -345,6 +356,7 @@ def test_pdf_get_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'format': 'json',
@@ -388,6 +400,7 @@ def test_get_prop_accesslog(app, model, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -430,6 +443,7 @@ def test_get_w_select_accesslog(app, model, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -471,6 +485,7 @@ def test_getall_accesslog(app, model, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -510,6 +525,7 @@ def test_getall_w_select_accesslog(app, model, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -558,6 +574,7 @@ def test_accesslog_file(model, postgresql, rc, request, tmp_path):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -668,6 +685,7 @@ def test_accesslog_file_stdin(
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'action': 'insert',
@@ -730,6 +748,7 @@ def test_accesslog_file_stderr(
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'action': 'insert',
@@ -773,6 +792,7 @@ def test_delete_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -813,6 +833,7 @@ def test_pdf_delete_accesslog(model, app, context):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -881,6 +902,7 @@ def test_pdf_ref_update_accesslog(model, app, context, tmp_path):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -932,6 +954,7 @@ def test_batch_write(model, app, context, tmp_path):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -977,6 +1000,7 @@ def test_stream_write(model, app, context, tmp_path):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -1022,6 +1046,7 @@ def test_ns_read(model, app, context, tmp_path):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -1066,6 +1091,7 @@ def test_ns_read_csv(model, app, context, tmp_path):
         {
             'txn': accesslog[-2]['txn'],
             'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
             'type': 'request',
             'time': accesslog[-2]['time'],
             'agent': 'testclient',
@@ -1083,5 +1109,181 @@ def test_ns_read_csv(model, app, context, tmp_path):
             'delta': accesslog[-1]['delta'],
             'memory': accesslog[-1]['memory'],
             'objects': objects[model],
+        },
+    ]
+
+
+@pytest.mark.manifests('internal_sql', 'csv')
+def test_get_accesslog_default_user(manifest_type: str, tmp_path: pathlib.Path, rc: RawConfig, postgresql: str, request: FixtureRequest):
+    context = bootstrap_manifest(
+        rc, '''
+        d | r | b | m | property      | type    | access
+        backends/postgres/dtypes/test |         | open
+          |   |   | Entity            |         |
+          |   |   |   | id            | integer |
+        ''',
+        backend=postgresql,
+        tmp_path=tmp_path,
+        manifest_type=manifest_type,
+        request=request,
+        full_load=True
+    )
+    context.set('accesslog.stream', [])
+    default_client_id = get_default_auth_client_id(context)
+    token = create_access_token(
+        context,
+        load_key(context, KeyType.private),
+        default_client_id,
+        int(datetime.timedelta(days=10).total_seconds()),
+        { 'spinta_getall', 'spinta_search'}
+    )
+
+    model = 'backends/postgres/dtypes/test/Entity'
+    app = create_test_client(context)
+    app.authmodel(model, ['insert'])
+
+    resp = app.post('/backends/postgres/dtypes/test/Entity', json={'id': 1})
+    assert resp.status_code == 201, resp.json()
+
+    # Get with default client auth
+    resp = app.get('/backends/postgres/dtypes/test/Entity?select(id)',  headers={'authorization': f'Bearer {token}'})
+    assert resp.status_code == 200, resp.json()
+
+    accesslog = context.get('accesslog.stream')
+    assert len(accesslog) == 4
+    assert accesslog[-2:] == [
+        {
+            'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
+            'type': 'request',
+            'time': accesslog[-2]['time'],
+            'client': default_client_id,
+            'method': 'GET',
+            'url': f'https://testserver/{model}?select(id)',
+            'agent': 'testclient',
+            'format': 'json',
+            'action': 'search',
+            'model': model,
+        },
+        {
+            'txn': accesslog[-2]['txn'],
+            'type': 'response',
+            'time': accesslog[-1]['time'],
+            'delta': accesslog[-1]['delta'],
+            'memory': accesslog[-1]['memory'],
+            'objects': 1,
+        },
+    ]
+
+
+@pytest.mark.manifests('internal_sql', 'csv')
+def test_get_accesslog_not_default_user(manifest_type: str, tmp_path: pathlib.Path, rc: RawConfig, postgresql: str,
+                                     request: FixtureRequest):
+    context = bootstrap_manifest(
+        rc, '''
+        d | r | b | m | property      | type    | access
+        backends/postgres/dtypes/test |         | open
+          |   |   | Entity            |         |
+          |   |   |   | id            | integer |
+        ''',
+        backend=postgresql,
+        tmp_path=tmp_path,
+        manifest_type=manifest_type,
+        request=request,
+        full_load=True
+    )
+    context.set('accesslog.stream', [])
+
+
+    model = 'backends/postgres/dtypes/test/Entity'
+    app = create_test_client(context)
+    app.authmodel(model, ['insert', 'getall', 'search'])
+
+    resp = app.post('/backends/postgres/dtypes/test/Entity', json={'id': 1})
+    assert resp.status_code == 201, resp.json()
+
+    # Get with test client
+    resp = app.get('/backends/postgres/dtypes/test/Entity?select(id)')
+    assert resp.status_code == 200, resp.json()
+
+    accesslog = context.get('accesslog.stream')
+    assert len(accesslog) == 4
+    assert accesslog[-2:] == [
+        {
+            'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
+            'scope': accesslog[-2]['scope'],
+            'type': 'request',
+            'time': accesslog[-2]['time'],
+            'client': 'test-client',
+            'method': 'GET',
+            'url': f'https://testserver/{model}?select(id)',
+            'agent': 'testclient',
+            'format': 'json',
+            'action': 'search',
+            'model': model,
+        },
+        {
+            'txn': accesslog[-2]['txn'],
+            'type': 'response',
+            'time': accesslog[-1]['time'],
+            'delta': accesslog[-1]['delta'],
+            'memory': accesslog[-1]['memory'],
+            'objects': 1,
+        },
+    ]
+
+
+@pytest.mark.manifests('internal_sql', 'csv')
+def test_get_accesslog_scope_log_false(manifest_type: str, tmp_path: pathlib.Path, rc: RawConfig, postgresql: str, request: FixtureRequest):
+    context = bootstrap_manifest(
+        rc, '''
+        d | r | b | m | property      | type    | access
+        backends/postgres/dtypes/test |         | open
+          |   |   | Entity            |         |
+          |   |   |   | id            | integer |
+        ''',
+        backend=postgresql,
+        tmp_path=tmp_path,
+        manifest_type=manifest_type,
+        request=request,
+        full_load=True
+    )
+    context.set('accesslog.stream', [])
+
+    # set config to not log scope
+    context.get('config').scope_log = False
+
+    model = 'backends/postgres/dtypes/test/Entity'
+    app = create_test_client(context)
+    app.authmodel(model, ['insert'])
+
+    resp = app.post('/backends/postgres/dtypes/test/Entity', json={'id': 1})
+    assert resp.status_code == 201, resp.json()
+
+    accesslog = context.get('accesslog.stream')
+    assert len(accesslog) == 2
+    assert accesslog[-2:] == [
+        {
+            'txn': accesslog[-2]['txn'],
+            'pid': accesslog[-2]['pid'],
+            'type': 'request',
+            'time': accesslog[-2]['time'],
+            'client': 'test-client',
+            'method': 'POST',
+            'url': f'https://testserver/{model}',
+            'agent': 'testclient',
+            'format': 'json',
+            'action': 'insert',
+            'model': model,
+            'rctype': 'application/json',
+        },
+        {
+            'txn': accesslog[-2]['txn'],
+            'type': 'response',
+            'time': accesslog[-1]['time'],
+            'delta': accesslog[-1]['delta'],
+            'memory': accesslog[-1]['memory'],
+            'objects': 1,
         },
     ]